+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Use black to enforce coding style
 - Migrate from requirements.txt to Pipfile
 - Use django-split-settings
+- Add search capabilities
+
 # 1.0.5 (2019-06-24)
 Fix plan datepicker